Break out selectors tests into individual files for Level 1 and Level 2
