+++ tests/test_api.py
@@ -522,69 +522,6 @@ class TestSoupSieve(util.TestCase):
 class TestInvalid(util.TestCase):
     """Test invalid."""
 
-    def test_invalid_pseudo_class_start_combinator(self):
-        """Test invalid start combinator in pseudo-classes other than `:has()`."""
-
-        self.assert_raises(':is(> div)', SyntaxError)
-        self.assert_raises(':is(div, > div)', SyntaxError)
-
-    def test_unrecognized_pseudo(self):
-        """Test unrecognized pseudo class."""
-
-        self.assert_raises(':before', NotImplementedError)
-
-    def test_invalid_pseudo_orphan_close(self):
-        """Test invalid, orphaned pseudo close."""
-
-        self.assert_raises('div)', SyntaxError)
-
-    def test_invalid_pseudo_dangling_comma(self):
-        """Test pseudo class group with trailing comma."""
-
-        self.assert_raises(':is(div,)', SyntaxError)
-
-    def test_invalid_pseudo_open(self):
-        """Test invalid pseudo close."""
-
-        self.assert_raises(':is(div', SyntaxError)
-
-    def test_invalid_incomplete_has(self):
-        """Test `:has()` fails with just a combinator."""
-
-        self.assert_raises(':has(>)', SyntaxError)
-
-    def test_invalid_has_empty(self):
-        """Test `:has()` fails with empty function parameters."""
-
-        self.assert_raises(':has()', SyntaxError)
-
-    def test_invalid_has_double_comma(self):
-        """Test `:has()` fails with consecutive commas."""
-
-        self.assert_raises(':has(> has,, a)', SyntaxError)
-
-    def test_invalid_has_double_combinator(self):
-        """Test `:has()` fails with consecutive combinators."""
-
-        self.assert_raises(':has(>> has a)', SyntaxError)
-        self.assert_raises(':has(> has, >> a)', SyntaxError)
-        self.assert_raises(':has(> has >> a)', SyntaxError)
-
-    def test_invalid_has_trailing_combinator(self):
-        """Test `:has()` fails with trailing combinator."""
-
-        self.assert_raises(':has(> has >)', SyntaxError)
-
-    def test_invalid_has_trailing_comma(self):
-        """Test `:has()` fails with trailing comma."""
-
-        self.assert_raises(':has(> has,)', SyntaxError)
-
-    def test_invalid_has_start_comma(self):
-        """Test `:has()` fails with trailing comma."""
-
-        self.assert_raises(':has(, p)', SyntaxError)
-
     def test_immutable_dict_hashable_value(self):
         """Test immutable dictionary has a hashable value."""
